slack-sdk-enrichment/requirements.txt
quixstreams<2.4
python-dotenv
slack_sdk